+++ spec/parallel_sftp/client_spec.rb
@@ -53,7 +53,7 @@ RSpec.describe ParallelSftp::Client do
       ).and_call_original
 
       expect(ParallelSftp::Download).to receive(:new)
-        .with(anything, on_progress: nil)
+        .with(anything, on_progress: nil, on_segment_progress: nil)
         .and_return(download_mock)
 
       client.download(remote_path, local_path)
@@ -86,10 +86,21 @@ RSpec.describe ParallelSftp::Client do
       progress_callback = ->(info) { puts info }
 
       expect(ParallelSftp::Download).to receive(:new)
-        .with(anything, on_progress: progress_callback)
+        .with(anything, on_progress: progress_callback, on_segment_progress: nil)
         .and_return(download_mock)
 
       client.download(remote_path, local_path, on_progress: progress_callback)
+    end
+
+    it "passes segment progress callback to Download" do
+      download_mock = instance_double(ParallelSftp::Download, execute: local_path)
+      segment_callback = ->(info) { puts info[:segments] }
+
+      expect(ParallelSftp::Download).to receive(:new)
+        .with(anything, on_progress: nil, on_segment_progress: segment_callback)
+        .and_return(download_mock)
+
+      client.download(remote_path, local_path, on_segment_progress: segment_callback)
     end
 
     it "returns the local path on success" do